Home work 7. MVP После отработки ДЗ - не добавил только интерфейс базовый. буду разбиратсья как это и что позже.
 - Не работает ktlint -( имена проверял ошибок не нашёл.
