--- app/src/main/java/ru/androidschool/intensiv/Domain/MovieEminClass.kt
@@ -1,7 +1,0 @@
-package ru.androidschool.intensiv.Domain
-
-enum class MovieEminClass(val tipeMovie: String) {
-    NAWPLAING("NawPlaing"),
-    UPCOMING("Upcomings"),
-    POPULAR("Popular")
-}
