--- app/src/main/java/ru/androidschool/intensiv/Presentation/EditTextExtensions.kt
@@ -1,8 +1,0 @@
-package ru.androidschool.intensiv.Presentation
-
-import android.text.Editable
-import android.widget.EditText
-import androidx.core.widget.addTextChangedListener
-
-fun EditText.afterTextChanged(action: (s: Editable?) -> Unit) =
-    addTextChangedListener(afterTextChanged = action)
