--- app/src/main/java/ru/androidschool/intensiv/Presentation/base/BasePresenter.kt
@@ -1,13 +1,0 @@
-package ru.androidschool.intensiv.Presentation.base
-
-abstract class BasePresenter<V> {
-    protected var view: V? = null
-
-    fun attachView(view: V) {
-        this.view = view
-    }
-
-    fun detachView() {
-        this.view = null
-    }
-}
