--- app/src/main/java/ru/androidschool/intensiv/Presentation/profile/ProfileAdapter.kt
@@ -1,19 +1,0 @@
-package ru.androidschool.intensiv.Presentation.profile
-
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import ru.androidschool.intensiv.Presentation.watchlist.WatchlistFragment
-
-class ProfileAdapter(
-    fragment: Fragment,
-    private val itemsCount: Int
-) : FragmentStateAdapter(fragment) {
-
-    override fun getItemCount(): Int {
-        return itemsCount
-    }
-
-    override fun createFragment(position: Int): Fragment {
-        return WatchlistFragment.newInstance()
-    }
-}
