--- app/src/main/java/ru/androidschool/intensiv/Presentation/profile/ProfileFragment.kt
@@ -1,87 +1,0 @@
-package ru.androidschool.intensiv.Presentation.profile
-
-import android.os.Bundle
-import android.text.SpannableString
-import android.text.style.RelativeSizeSpan
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.widget.ViewPager2
-import com.google.android.material.tabs.TabLayoutMediator
-import com.squareup.picasso.Picasso
-import jp.wasabeef.picasso.transformations.CropCircleTransformation
-import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.databinding.FragmentProfileBinding
-
-class ProfileFragment : Fragment() {
-
-    private lateinit var profileTabLayoutTitles: Array<String>
-
-    private var _binding: FragmentProfileBinding? = null
-
-    // This property is only valid between onCreateView and
-    // onDestroyView.
-    private val binding get() = _binding!!
-
-    private var profilePageChangeCallback = object : ViewPager2.OnPageChangeCallback() {
-        override fun onPageSelected(position: Int) {
-            Toast.makeText(
-                requireContext(),
-                "Selected position: $position",
-                Toast.LENGTH_SHORT
-            ).show()
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentProfileBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        Picasso.get()
-            .load(R.drawable.ic_avatar)
-            .transform(CropCircleTransformation())
-            .placeholder(R.drawable.ic_avatar)
-            .into(binding.avatar)
-
-        profileTabLayoutTitles = resources.getStringArray(R.array.tab_titles)
-
-        val profileAdapter = ProfileAdapter(
-            this,
-            profileTabLayoutTitles.size
-        )
-        binding.profileViewPager.adapter = profileAdapter
-
-        binding.profileViewPager.registerOnPageChangeCallback(
-            profilePageChangeCallback
-        )
-
-        TabLayoutMediator(binding.tabLayout, binding.profileViewPager) { tab, position ->
-
-            // Выделение первой части заголовка таба
-            // Название таба
-            val title = profileTabLayoutTitles[position]
-            // Раздеряем название на части. Первый элемент будет кол-во
-            val parts = profileTabLayoutTitles[position].split(" ")
-            val number = parts[0]
-            val spannableStringTitle = SpannableString(title)
-            spannableStringTitle.setSpan(RelativeSizeSpan(2f), 0, number.count(), 0)
-
-            tab.text = spannableStringTitle
-        }.attach()
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
