--- app/src/main/java/ru/androidschool/intensiv/Presentation/watchlist/MoviePreviewItem.kt
@@ -1,30 +1,0 @@
-package ru.androidschool.intensiv.Presentation.watchlist
-
-import android.view.View
-import com.squareup.picasso.Picasso
-import com.xwray.groupie.viewbinding.BindableItem
-import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.dto.MovieEntity
-import ru.androidschool.intensiv.databinding.ItemSmallBinding
-
-class MoviePreviewItem(
-    private val content: MovieEntity,
-    private val onClick: (movie: MovieEntity) -> Unit
-) : BindableItem<ItemSmallBinding>() {
-
-    override fun getLayout() = R.layout.item_small
-
-    override fun bind(view: ItemSmallBinding, position: Int) {
-        view.imagePreview.setOnClickListener {
-            onClick.invoke(content)
-        }
-        // TODO Получаем данные из базы
-        Picasso.get()
-            .load(content.posterPath)
-            .fit()
-            .centerCrop()
-            .into(view.imagePreview)
-    }
-
-    override fun initializeViewBinding(v: View): ItemSmallBinding = ItemSmallBinding.bind(v)
-}
