+++ app/src/main/java/ru/androidschool/intensiv/data/roomdata/MovieDatabase.kt
@@ -1,0 +1,24 @@
+package ru.androidschool.intensiv.data.roomdata
+
+import android.content.Context
+import androidx.room.*
+
+@Database(entities = [MovieEntity::class], version = 1)
+abstract class MovieDatabase : RoomDatabase() {
+    abstract fun movieDao(): MovieDao
+
+    companion object {
+        private var instance: MovieDatabase? = null
+
+        @Synchronized
+        fun get(context: Context): MovieDatabase {
+            if (instance == null) {
+                instance = Room.databaseBuilder(
+                    context.applicationContext,
+                    MovieDatabase::class.java, "MovieDatabase"
+                ).build()
+            }
+            return instance!!
+        }
+    }
+}
