+++ app/src/main/java/ru/androidschool/intensiv/domain/MovieEminClass.kt
@@ -1,0 +1,7 @@
+package ru.androidschool.intensiv.domain
+
+enum class MovieEminClass(val tipeMovie: String) {
+    NOWPLAYING("NawPlaying"),
+    UPCOMING("Upcomings"),
+    POPULAR("Popular")
+}
