+++ app/src/main/java/ru/androidschool/intensiv/domain/extensionFunk.kt
