+++ app/src/main/java/ru/androidschool/intensiv/domain/repository/MovieDomainRepository.kt
@@ -1,0 +1,9 @@
+package ru.androidschool.intensiv.domain.repository
+
+import io.reactivex.Single
+
+import ru.androidschool.intensiv.data.vo.MovieVo
+
+interface MovieDomainRepository {
+    fun getMovies(): Single<List<MovieVo>>
+}
