+++ app/src/main/java/ru/androidschool/intensiv/presentation/SearchBar.kt
@@ -1,0 +1,61 @@
+package ru.androidschool.intensiv.presentation
+
+import android.content.Context
+import android.util.AttributeSet
+import android.view.LayoutInflater
+import android.view.View
+import android.widget.FrameLayout
+import androidx.core.view.isVisible
+import ru.androidschool.intensiv.R
+import ru.androidschool.intensiv.databinding.SearchToolbarBinding
+
+class SearchBar @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyle: Int = 0
+) : FrameLayout(context, attrs, defStyle) {
+
+    lateinit var binding: SearchToolbarBinding
+
+    private var hint: String = ""
+    private var isCancelVisible: Boolean = true
+
+    init {
+        if (attrs != null) {
+            context.obtainStyledAttributes(attrs, R.styleable.SearchBar).apply {
+                hint = getString(R.styleable.SearchBar_hint).orEmpty()
+                isCancelVisible = getBoolean(R.styleable.SearchBar_cancel_visible, true)
+                recycle()
+            }
+        }
+    }
+
+    fun setText(text: String?) {
+        binding.searchEditText.setText(text)
+    }
+
+    fun clear() {
+        binding.searchEditText.setText("")
+    }
+
+    override fun onFinishInflate() {
+        super.onFinishInflate()
+        binding = SearchToolbarBinding.inflate(LayoutInflater.from(context), this, true)
+        binding.searchEditText.hint = hint
+        binding.deleteTextButton.setOnClickListener {
+            binding.searchEditText.text.clear()
+        }
+    }
+
+    override fun onAttachedToWindow() {
+        super.onAttachedToWindow()
+        binding.searchEditText.afterTextChanged { text ->
+            if (!text.isNullOrEmpty() && !binding.deleteTextButton.isVisible) {
+                binding.deleteTextButton.visibility = View.VISIBLE
+            }
+            if (text.isNullOrEmpty() && binding.deleteTextButton.isVisible) {
+                binding.deleteTextButton.visibility = View.GONE
+            }
+        }
+    }
+}
