+++ app/src/main/java/ru/androidschool/intensiv/presentation/base/BasePresenter.kt
@@ -1,0 +1,13 @@
+package ru.androidschool.intensiv.presentation.base
+
+abstract class BasePresenter<V> {
+    protected var view: V? = null
+
+    fun attachView(view: V) {
+        this.view = view
+    }
+
+    fun detachView() {
+        this.view = null
+    }
+}
