+++ app/src/main/java/ru/androidschool/intensiv/presentation/feed/MovieItem.kt
@@ -1,0 +1,33 @@
+package ru.androidschool.intensiv.presentation.feed
+
+import android.view.View
+import com.squareup.picasso.Picasso
+import com.xwray.groupie.viewbinding.BindableItem
+import ru.androidschool.intensiv.R
+import ru.androidschool.intensiv.data.dto.Movie
+import ru.androidschool.intensiv.databinding.ItemWithTextBinding
+
+class MovieItem(
+    private val content: Movie,
+    private val onClick: (movie: Movie) -> Unit
+) : BindableItem<ItemWithTextBinding>() {
+
+    override fun getLayout(): Int = R.layout.item_with_text
+
+    override fun bind(view: ItemWithTextBinding, position: Int) {
+        view.description.text = content.title
+        view.movieRating.rating = content.rating
+        view.content.setOnClickListener {
+            onClick.invoke(content)
+        }
+
+        // TODO Получать из модели
+        Picasso.get()
+            .load(content.posterPath)
+            .fit()
+            .centerCrop()
+            .into(view.imagePreview)
+    }
+
+    override fun initializeViewBinding(v: View) = ItemWithTextBinding.bind(v)
+}
