+++ app/src/main/java/ru/androidschool/intensiv/presentation/profile/ProfileAdapter.kt
@@ -1,0 +1,19 @@
+package ru.androidschool.intensiv.presentation.profile
+
+import androidx.fragment.app.Fragment
+import androidx.viewpager2.adapter.FragmentStateAdapter
+import ru.androidschool.intensiv.presentation.watchlist.WatchlistFragment
+
+class ProfileAdapter(
+    fragment: Fragment,
+    private val itemsCount: Int
+) : FragmentStateAdapter(fragment) {
+
+    override fun getItemCount(): Int {
+        return itemsCount
+    }
+
+    override fun createFragment(position: Int): Fragment {
+        return WatchlistFragment.newInstance()
+    }
+}
