+++ app/src/main/res/layout/tv_shows_fragment.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="ru.androidschool.intensiv.Presentation.tvshows.TvShowsFragment">
+    tools:context="ru.androidschool.intensiv.presentation.tvshows.TvShowsFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
